Improve Chatting Functionality
Client can now chat with the server with more than just an echo, and
does so in an async manner! Receives messages while typing the last
message!

TODO: Right now you have to ctrl+c to exit, allow it to properly exit.
TODO: Allow user to enter a name for the chat.